re-organize code in device_plugin_manager for better code visibility rights
